got it to a point where we can run a python server which broadcasts incoming messages from clients to all other clients
next we can probably start on deciding a format for data to be encoded in (this would be particularly useful if we want to 'stream' the state of our board to the other board - also for this, we may need to instead use UDP if TCP is too slow - will try to implement this with TCP first, and see how this goes

after we have a format for messages, we can develop the actual game (at least for 1 board), and the integrate message sending over wifi. may want to look into structuring this nicely as ideally everything wifi related is hidden from the game logic.

might want to extend the server after the game is working to store a game state/history permanently, and allow for some kind of scoreboard and match replays through a frontend